rename: 게시글 목록 파일들 보더 파일 내부로 이동(#121)
