+++ src/router/AppRouter.tsx
@@ -46,9 +46,9 @@ import TestSurveyList from "@src/pages/test/BoardList/TestSurveyList";
 import TestJobBannerPage from "@src/pages/test/TestJobBanner";
 
 import { isValidOzKeyLocal, resolveCohortNumber } from "@src/utils/auth/ozKey";
-import PostDetailPage from "@src/pages/PostDetailPage";
-import PostListPage from "@src/pages/PostListPage";
-import SurveyListPage from "@src/pages/SurveyListPage";
+import PostDetailPage from "@src/pages/boards/PostDetailPage";
+import PostListPage from "@src/pages/boards/PostListPage";
+import SurveyListPage from "@src/pages/boards/SurveyListPage";
 
 import TestGithubList from "@src/pages/test/BoardList/TestGithubList";
 import MainLayout from "@src/layouts/MainLayout";
@@ -56,7 +56,6 @@ import FreeBoard from "@src/pages/boards/FreeBoard";
 import JobsBoard from "@src/pages/boards/JobsBoard";
 import InfoBoard from "@src/pages/boards/InfoBoard";
 import WritePostPage from "@src/components/Board/WritePostPage";
-
 
 /** 공개(보호 불필요) 경로 목록 */
 const PUBLIC_PATHS: ReadonlySet<string> = new Set([
@@ -121,7 +120,7 @@ function KeyVerifyPlaceholder() {
     } catch (err: unknown) {
       const msg =
         typeof err === "object" && err && "message" in err
-          ? ((err as { message?: string }).message ?? "인증 실패")
+          ? (err as { message?: string }).message ?? "인증 실패"
           : "인증 실패";
       push({ message: msg, type: "error" });
     }
